+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.14.2] - 2024-08-21
+
+### Removed
+
+- **[ENGINE BC]** Removed generic `fixtures.TestCommand`, `TestEvent`,
+  `TestTimeout` and related `TypeA` through `TypeZ` types.
+
 ## [0.14.1] - 2024-08-18
 
 ### Deprecated
@@ -346,6 +353,7 @@ No engines except [testkit] are able to provide a meaningful implementation of
 [0.13.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.13.1
 [0.14.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.14.0
 [0.14.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.14.1
+[0.14.2]: https://github.com/dogmatiq/dogma/releases/tag/v0.14.2
 
 <!-- adr references -->